This should work but it does not ;(
I implement the solution as a simple sum over the overlapping pieces.
For each claim I add its pieces to the board and check the size of the
board and compare it to the expected size of the board (sum of the
pieces to be added with the currebt size of the board). I add the the
difference of those two quantities.